requirements.txt
pytest==7.2.1
pytest-clarity
mypy
curlify==2.2.1
pytest-xdist==3.0.2
allure-pytest==2.9.45